Update ReadMe.md
In line 13 I have generated a new line of note, which describes about the use of the stuff, because before it generating a line space between the bullets and was also looking something ugly. 
Another at line 80, I have a code tag which highlights the path or structure for the apk.